Created prerelease-windows.bat and fixed a number of small issues.
