+++ .gitignore
@@ -1,3 +1,4 @@
+AimTestProject
 *.exp
 
 # Created by https://www.toptal.com/developers/gitignore/api/python,c++,ninja
